Moved querytype package from api to internal.
