+++ src/main/java/org/mule/extension/elastic/internal/operations/SearchOperations.java
@@ -25,10 +25,10 @@ import org.elasticsearch.search.builder.SearchSourceBuilder;
 import org.mule.extension.elastic.api.JsonData;
 import org.mule.extension.elastic.api.SearchRequestConfiguration;
 import org.mule.extension.elastic.api.SearchSourceConfiguration;
-import org.mule.extension.elastic.api.querytype.Query;
 import org.mule.extension.elastic.internal.connection.ElasticsearchConnection;
 import org.mule.extension.elastic.internal.error.ElasticsearchError;
 import org.mule.extension.elastic.internal.error.exception.ElasticsearchException;
+import org.mule.extension.elastic.internal.querytype.Query;
 import org.mule.extension.elastic.internal.utils.ElasticsearchUtils;
 import org.mule.runtime.extension.api.annotation.param.Connection;
 import org.mule.runtime.extension.api.annotation.param.MediaType;
