+++ src/main/java/org/mule/extension/elastic/internal/querytype/BaseQueryString.java
@@ -1,7 +1,7 @@
 /**
  * Copyright (c) 2003-2017, Great Software Laboratory Pvt. Ltd. The software in this package is published under the terms of the Commercial Free Software license V.1, a copy of which has been included with this distribution in the LICENSE.md file.
  */
-package org.mule.extension.elastic.api.querytype;
+package org.mule.extension.elastic.internal.querytype;
 
 import java.util.Map;
